version 1.0 - Prete pr front end

+++ src/main/java/com/srda/mpam/model/dto/OperationDto.java
@@ -13,7 +13,7 @@ import java.time.LocalDate;
 public class OperationDto {
 
     private Long id;
-    private Long montant;
+    private Double montant;
     private LocalDate dateValeur;
     private Long refBanquaire;
     private Long accountUserId;
